\name{twoe-package}
\alias{twoe-package}
\alias{twoe}
\docType{package}
\title{
Software for modelling tropical forest dynamics
}
\description{
"twoe" is a software which aims first at estimating demographic parameters for tropical tree species from permanent forest plot data (through the "entry" module) and second at simulating forest dynamics (through the "exit" module).
}
\details{
\tabular{ll}{
Package: \tab twoe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-08\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}
\author{
Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>
}
\keyword{tropical forest dynamics}
